Provide means to avoid SparkPost's tracking option
With the spark post API, it's possible to disable tracking
information using ->setOptions(). Sometimes some users don't
want these tracking information within their email content so
add support in this extension to configure this behavior.

This introduces 3 config variables that can be used in the
LocalSettings.php file.
- $wgSparkpostOpenTracking
- $wgSparkpostClickTracking
- $wgSparkpostTransactional

These can be set to false or true depending on the case but will
default to "true" if not set as this is normal behavior in spark
post.

Bug: T214664
Change-Id: I81a2b8315ba6c2b267fa4b241ce7bab05724c2bc